MNT Fix link test to not pre-cache link owner before test starts. (#392)
With the new caching, this will fail tests that add versioned to the
class. That's because of a combination of factors:
- extension instances are stored on instances instead of accessed
  statically
- in `DataObject::get_one()`, the `getUniqueKeyComponents()` method is
  used to determine the cache key, which just _happens_ to be called on
a singleton, and we just _happen_ to have not generated a singleton of
the class until now, which means it has the versioned extension freshly
added to it.

This would be extremely unlikely to occur in real code, and I think our
response to a bug report would be "reset your cache after applying the
extension, or just don't fiddle around with extensions at runtime except
inside `_config.php`".